# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartCity OS Design Tools API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /snapshots:
    get:
      operationId: listSnapshots
      tags: [snapshots]
      summary: List Revit snapshots across all engagements
      description: Returns all snapshots, newest first, joined with engagementName.
      responses:
        "200":
          description: Array of snapshot summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotSummary"
    post:
      operationId: createSnapshot
      tags: [snapshots]
      summary: Receive a Revit snapshot from the add-in
      description: |
        Stores a snapshot from the Revit add-in. Requires the `x-snapshot-secret`
        header to match the configured `SNAPSHOT_SECRET` env var. If `projectName`
        matches an existing engagement (case-insensitive), the snapshot attaches
        to it. Otherwise a new engagement is auto-created.
      parameters:
        - name: x-snapshot-secret
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotPayload"
      responses:
        "201":
          description: Snapshot accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotReceipt"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid snapshot secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /snapshots/{id}:
    get:
      operationId: getSnapshot
      tags: [snapshots]
      summary: Get the full snapshot record (with payload)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SnapshotSummary:
      type: object
      properties:
        id:
          type: string
        engagementId:
          type: string
        engagementName:
          type: string
        projectName:
          type: string
        sheetCount:
          type: integer
          nullable: true
        roomCount:
          type: integer
          nullable: true
        levelCount:
          type: integer
          nullable: true
        wallCount:
          type: integer
          nullable: true
        receivedAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - engagementName
        - projectName
        - sheetCount
        - roomCount
        - levelCount
        - wallCount
        - receivedAt
    SnapshotPayload:
      type: object
      properties:
        projectName:
          type: string
      required:
        - projectName
      additionalProperties: true
    SnapshotReceipt:
      type: object
      properties:
        id:
          type: string
        receivedAt:
          type: string
          format: date-time
        engagementId:
          type: string
        engagementName:
          type: string
        autoCreated:
          type: boolean
      required:
        - id
        - receivedAt
        - engagementId
        - engagementName
        - autoCreated
    SnapshotDetail:
      type: object
      properties:
        id:
          type: string
        engagementId:
          type: string
        engagementName:
          type: string
        projectName:
          type: string
        sheetCount:
          type: integer
          nullable: true
        roomCount:
          type: integer
          nullable: true
        levelCount:
          type: integer
          nullable: true
        wallCount:
          type: integer
          nullable: true
        receivedAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
        sheets:
          type: array
          items:
            $ref: "#/components/schemas/SheetSummary"
      required:
        - id
        - engagementId
        - engagementName
        - projectName
        - sheetCount
        - roomCount
        - levelCount
        - wallCount
        - receivedAt
        - payload
        - sheets
    SheetSummary:
      type: object
      properties:
        id:
          type: string
        snapshotId:
          type: string
        engagementId:
          type: string
        sheetNumber:
          type: string
        sheetName:
          type: string
        viewCount:
          type: integer
          nullable: true
        revisionNumber:
          type: string
          nullable: true
        revisionDate:
          type: string
          nullable: true
        thumbnailWidth:
          type: integer
        thumbnailHeight:
          type: integer
        fullWidth:
          type: integer
        fullHeight:
          type: integer
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - snapshotId
        - engagementId
        - sheetNumber
        - sheetName
        - viewCount
        - revisionNumber
        - revisionDate
        - thumbnailWidth
        - thumbnailHeight
        - fullWidth
        - fullHeight
        - sortOrder
        - createdAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...